// src/constants/types/dataShopType.ts
export interface DataShopTypeInterface {
  id: string;
  categoryId: string;
  offer: string;
  price: number;
  nameEn: string;
  nameEs: string;
  nameDe: string;
  nameIt: string;
  nameFr: string;
  nameRu: string;
  path: string;
  pathToPost: string;
  type: string;
  secondType: string;
  keyWords: string;
  dateCreate: number;
  totalPurchased: number;
  inStock: number;
  like: number;
  productName: string;
  image: string;
  descriptionsDe: string;
  descriptionsEn: string;
  descriptionsEs: string;
  descriptionsIt: string;
  descriptionsFr: string;
  descriptionsRu: string;
}

export interface SearchShopTypeInterface {
  id: string;
  categoryId: string;
  offer: string;
  price: number;
  nameEn: string;
  nameEs: string;
  nameDe: string;
  nameIt: string;
  nameFr: string;
  nameRu: string;
  path: string;
  type: string;
  pathToPost: string;
  secondType: string;
  keyWords: string;
  dateCreate: number;
  totalPurchased: number;
  inStock: number;
  like: number;
  productName: string;
  image: string;
  descriptionsDe: string;
  descriptionsEn: string;
  descriptionsEs: string;
  descriptionsIt: string;
  descriptionsFr: string;
  descriptionsRu: string;
}